+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Rolling all the math commands up into one "rasterman math" command.
 * Re-implement the area thresholding as a non-recursive algorithm. 
 * **New Operation**: `stackstats`: Create a raster with the cumulative stats for each cell in a stack of rasters.
+* **New Operation**: `areathresh`: Threshold out small areas delineated by nodata values. 
+* **New Operation**: `smoothedges`: subtract cells from the edge and add them back to smooth things.
 
 ### Fixes